Add Respoke-SDK header to requests
This patch adds the "Respoke-SDK" header containing the library
version, rough operating system version, and the .net CLR version.
I was disappointed to find that it is very difficult to find the
.net platform version that the app is running, so the CLR version
was the next best thing. The operating system that is reported is a
very rough value. On Mac, it reports "Unix". However, it's good enough
for now and if we decide we need more granular information later
we can improve on it.

Related to MER-4340.